@@ -2,40 +2,36 @@
   "name": "translation-portal-react",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "homepage": "https://translate.sunbird.ai",
   "dependencies": {
     "@emotion/core": "^11.0.0",
-    "@emotion/react": "^11.8.2",
-    "@emotion/styled": "^11.8.1",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.0",
     "@hotjar/browser": "^1.0.9",
-    "@mui/icons-material": "^5.6.0",
-    "@mui/material": "^5.5.2",
-    "@testing-library/jest-dom": "^5.16.3",
-    "@testing-library/react": "^12.1.4",
-    "@testing-library/user-event": "^13.5.0",
-    "gh-pages": "^3.2.3",
-    "p-retry": "^5.1.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-ga4": "^1.4.1",
-    "styled-components": "^6.1.11",
-    "twin.macro": "^2.8.2",
+    "@mui/icons-material": "^6.4.6",
+    "@mui/material": "^6.4.6",
+    "gh-pages": "^6.3.0",
+    "globals": "^16.0.0",
+    "p-retry": "^6.2.1",
+    "react": "^18",
+    "react-dom": "^18",
+    "react-ga4": "^2.1.0",
+    "styled-components": "^6.1.15",
+    "tailwindcss": "^3.3.3",
+    "twin.macro": "^3.4.1",
     "typewriter-effect": "^2.21.0",
-    "web-vitals": "^2.1.4"
+    "vite": "^6.2.0",
+    "vite-plugin-babel-macros": "^1.0.6",
+    "web-vitals": "^4.2.4"
   },
   "scripts": {
+    "dev": "vite",
     "predeploy": "npm run build",
     "deploy": "gh-pages -d build",
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
+    "start": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
   },
   "browserslist": {
     "production": [
@@ -50,10 +46,13 @@
     ]
   },
   "devDependencies": {
-    "autoprefixer": "^10.4.4",
-    "cypress": "^13.14.0",
-    "postcss": "^8.4.12",
-    "react-scripts": "^5.0.1",
-    "tailwindcss": "^3.0.23"
+    "@eslint/js": "^9.21.0",
+    "@vitejs/plugin-react": "^4.3.4",
+    "autoprefixer": "^10.4.20",
+    "babel-plugin-macros": "^3.1.0",
+    "cypress": "^14.1.0",
+    "eslint": "^9.21.0",
+    "eslint-plugin-react": "^7.37.4",
+    "vite-plugin-eslint": "^1.8.1"
   }
 }
