@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
@@ -10,11 +10,16 @@ const hotjarVersion = 6;
 
 Hotjar.init(siteId, hotjarVersion);
 
-ReactDOM.render(
+// Create root once
+const root = ReactDOM.createRoot(
+  document.getElementById("root") 
+);
+
+// Then render to that root
+root.render(
   <React.StrictMode>
     <App />
-  </React.StrictMode>,
-  document.getElementById('root')
+  </React.StrictMode>
 );
 
 // If you want to start measuring performance in your app, pass a function
