@@ -1,7 +1,7 @@
 import ReactGA from "react-ga4";
 
 export const TrackGoogleAnalyticsEvent = (category, action, label) => {
-  if (process.env.REACT_APP_NODE_ENV === "Production") {
+  if (import.meta.env.MODE === "Production") {
     if (!category || !action) {
       console.error("trackEvent: category and action are required parameters");
       return;
